requirements.txt
streamlit
openai
pandas
numpy
st-annotated-text
streamlit-aggrid
